+++ src/views/UserSearch/UserSearch.js
@@ -582,13 +582,6 @@ class UserSearch extends React.Component {
   }
 
   handleFinancialTransactionsReportFormSubmit = async (data) => {
-    const {
-      startDate,
-      endDate,
-      servicePoint,
-      feeFineOwner,
-    } = data;
-
     if (this.state.financialTransactionsReportInProgress) {
       return;
     }
@@ -599,7 +592,15 @@ class UserSearch extends React.Component {
     });
 
     const {
+      startDate,
+      endDate,
+      servicePoint,
+      feeFineOwner,
+    } = data;
+    const {
+      resources,
       mutator: { financialTransactionsReport },
+      intl,
     } = this.props;
     const reportParameters = {
       createdAt: servicePoint.map(s => s.value),
@@ -618,9 +619,11 @@ class UserSearch extends React.Component {
           message: <FormattedMessage id="ui-users.reports.noItemsFound" />,
         });
       } else {
+        const selectedOwner = resources.owners.records.find(({ id }) => id === feeFineOwner);
         const headerData = {
           ...reportParameters,
           createdAt: reportParameters.createdAt.join(', '),
+          feeFineOwner: selectedOwner.owner,
         };
         const reportParams = {
           data: reportData,